Added 2 new Values methods for wanted latlon-matrix calculations: 1. Basic method for current time 2. Second method for time-interpolations with theTimeRangeInMinutes and doNearestTimeIfPossible parameters
Fixed InterpolationHelper (which is for time-interpolations) so that if either of given values is missing, check from theFactor1 which value is returned (the closer to the wanted time). This used to return the value non-missing, which is different than is usually done with time-interpolations else where.
This change effects 4 matrix calculation methods:
1. NFmiFastQueryInfo::Values(NFmiDataMatrix<float> &theMatrix, const NFmiMetTime &theInterpolatedTime)
2. void NFmiFastQueryInfo::CroppedValues(NFmiDataMatrix<float> &theMatrix, const NFmiMetTime &theInterpolatedTime,...)
3. void NFmiFastQueryInfo::LandscapeValues(NFmiDataMatrix<float> &theMatrix, const NFmiMetTime &theInterpolatedTime,...)
4. NFmiFastQueryInfo::LandscapeCachedInterpolation(...)