Added new CMake option in comments: set CMAKE_SYSTEM_VERSION to 10.0
I found out that Visual C++ 2015 acts strangely with Add member variable wizard when working on Windows 10 platform, when project uses Windows 8.1 SDK as target system. So you better use same target as you are working on, but I couldn't set to 10 because I don't have the SDK for Windows 10 and I didn't want to make this big change here.

In future we should change to Visual C++ 2017 and Windows 10.0 SDK.