CaseStudy system had one big flaw when selecting the stored data: You couldn't skip latest data and select e.g. yesterdays older data alone. Changed "CS Count" (= how many latest data will be stored) to "CS Index-Range" column in the CaseStudy-dialog. This index-range is used following way: 1. Value 3-1 means: store all 3 latest values of model data, index 1 means the latest and 2 means the 2nd latest and 3 means the 3rd latest data. 2. Value 1-1 means that just the latest data will be stored, which is fixed value for observations and analyse. 3. If you want to select e.g. yesterdays data by skipping couple of latest data, something like 4-3 index-range will do that, only 3rd and 4th latest data will be stored.
Also added help into tooltips about which model-data will be stored and which not. There will be list of data files with their indices (starting from 1) and there will be each data's model-run-times and those data files that are about to be stored will be drawn with green text and those that are not will be drawn with red text.

Added new CaseStudyIndex1 and CaseStudyIndex2 sections to the windows registry to store this new index-range concept values.