Original problem: In SmartMet's macroParam calculation when a lot of observation parameters were used, execution times were really slow.
Root reason for the problem:
newbase's querydata iterator system (quryInfo, fastInfo, etc.) was designed poorly and when doing shallow copies of iterators from the same queryData, the data is not copied, but e.g. all the stations in locationBag are copied. In case of Synop data, there are around 10000 stations that are copied for each used synop parameter in the macroParam script, and this happened multiple times for each parameter depending of circumstances.

Way to optimize fastInfo iterator copy:
Added new feature to NFmiLocationBag that allows user to put locationBag object into optimization mode.
Also put itsLocations, itsSortedLocations and itsNearTree objects in to shared_ptr, that allows actual optimization code.
When in this mode copy constructor and assignment operator= just do shared_ptr copy.
When not in this mode, the old way is used by adding each new location copy into those 3 structures.

Problem with this optimization:
Newbase should be redesigned so that these shallow copies were used normally (this won't happen any more and this should have been done in the 90's). This optimization can't be used as default, because there are cases when using locationBag objects would break existing codes.
I tried to test as thoroughly as possible, but still wondering what went wrong and when these problems will surface.
For these reasons I won't merge these changes into 5.14 wgs84 development branch, so that nothing on Linux side don't break.